pyproject.toml
[project]
name = "ghost-code"
version = "0.1.0"
description = "動画への不可視フォレンジック透かしシステム"
requires-python = ">=3.11,<3.12"
dependencies = [
    "numpy>=1.24,<2.0",
    "pillow>=10.0",
    "opencv-python>=4.10",
    "torch",
    "torchvision",
    "trustmark>=0.7",
    "reedsolo>=1.7",
    "ffmpeg-python>=0.2",
    "tqdm>=4.66",
    "gradio>=6.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "ipykernel>=6.29",
]

[tool.uv]
# PyTorch は CUDA 12.8(Blackwell / RTX 50系)対応版を別 index から取得
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ghost_code"]
